rethink Drag a bit
still think Drag, Zoom, other event handlers should all be multiplexed into a single On function on the interpreter typeclass OR added as extensions to the interpreter. The latter looks like using features of Tagless just for the sake of it, from this vantage point so we’ll aim for the former.

going to implement some mouse events, paralleling exactly this drag addition and then try and collapse all the event handlers back down after that