WIP commit - removes old svg on each story
this is plumbed in in the most direct (ie repetitive) manner imaginable. it should be done by the higher level component but first might want to just check whether it’s desirable to re-use the SVGs in some ways, ie in the tree example.
whether for transitions or some other reason. so the re-factor is deferred