working on the GUP page
i think this might be a good direction for a working method as far as building on Ocelot goes - by keeping the Ocelot fork as a submodule one can prototype with reference to the Blocks in the Ocelot sources but then copy and customise them in the application later. That’s what this commit does with regard to the Card block. Also included here are some proofs of concept for leaning more on tailwind @apply in the test-apply.css than on spelling things out in the PureScript. There’s a couple of reasons for this and one of them is with an eye to the JIT compilation of CSS bundle with tailwind in the future.