WIP commit moving to better model of forces
introduced in this commit is the idea that the `datum_` spec in the model should expose filtering / discriminating selectors such as “isPackage” or “isUnusedModule” which will enable Forces to be configured and reconfigured at a higher level in the DSL.

Also, the idea of the “Fixed Force” is introduced - under the hood this is not a D3 force because it isn’t used by the simulation engine (except in the sense that the simulation engine doesn’t touch nodes which have `fx` / `fy` set). Again, the idea here is to make pinning nodes behave like a Force rather than having some orthogonal controls in that regard.

note that some of this is coming out of deeper thinking of how a higher level DSL might work and some is coming directly from UI/UX requirements from Spago. Hopefully the Spago app requirements generalize somewhat.