re-factor how filtering of forces works
not finished in this commit but the idea is that a filter function MAYBE present in the Force and if it is then the strength calculation will only apply to nodes that match the filter
in other words the filter function will be wrapped around the Attr for strength. somehow.