fix GUP
so now some of the D3 complexity has crept back in to this simple example, we have to open a selection everytime in the lambda that we return from the script.

this is essentially exposing the .enter() part of the D3 pattern where we don’t really want it in order to support the more complex examples involving simulations

possibly this is an argument for bringing back the simpler version of the update join?