nodeInitializer stuff added
so, there's an array of functions that's folded over the filtered-from-the-model nodes before they're put into staging. this allows gridXY (or anything else that you need to calculate on the staging data to be added - if you can calculate this upfront in the model then obviously you'd do it there but here we're concerned with things that depend on what actually makes it into the staging after filtering and functions that depend on what it is that has made it into staging).

a good example of the sort of thing that could be done here is responsive grid size depending on the numbers of objects that are being gridded. Or initializing xy to NaN so that D3 will do the phylotaxis thing or putting the tree nodes at their calculated treeXY so that they don't have so far to move when the forces act on them etc etc etc