WIP commit before rejigging order in Draw
There's a problem with initialPositioning being set to NaN so that D3 will layout the nodes in phyllotaxis...and that is that the translate attribute is getting set before the nodes are put in the simulation.

after they get put in the simulation the d.x is no longer NaN

the question is - is it possible to the attribute setting after putting the nodes in the simulation? how early can it be done? this commit is to clear the decks to try various ideas here