add setNodesFromSelection/setLinksFromSelection
this is less type safe than the setNodes / setLinks route but with D3 as a backend these at least hide the unsafeCoerce needed to go from the untyped data returned by the selection to something we can put in the simulation.

NB these are only needed if you are doing updating, simulation running, continutity-needed things like the Spago example. For less complicated examples, like LesMis there's no need for it

There may be a clever way - something like purescript-exists to encode recovery of the types type-safely