add a renderNotebook_ for simpler cases
if you don't need to provide state (actually substate) to any of the cells as in the ThreeLittleCircles examples then we can drop that state parameter.

undecided whether it's worth providing simpler types too? i think it's okay that you specify Unit in the signature but don't have to pass unit to the renderNotebook_ function