preparatory work on Spago app to generalize it
It can be seen that the front page “force layout navigator” has serious problems around the setup and management of callbacks, and needs to be rebuilt as a close copy of Spago.  however it’s probably worth taking a moment to generalize the Spago example before doing this. This commit makes one such step, removing the model reading code and associated Aff stuff from the State module